refector query code
