+++ src/main/java/de/samply/bbmri/negotiator/control/ApplicationBean.java
@@ -43,6 +43,7 @@ import de.samply.bbmri.negotiator.db.util.DbUtil;
 import de.samply.bbmri.negotiator.jooq.tables.pojos.Query;
 import de.samply.bbmri.negotiator.jooq.tables.records.QueryRecord;
 import eu.bbmri.eric.csit.service.negotiator.database.DbUtilCollection;
+import eu.bbmri.eric.csit.service.negotiator.database.DbUtilQuery;
 import eu.bbmri.eric.csit.service.negotiator.database.DbUtilRequest;
 import eu.bbmri.eric.csit.service.negotiator.lifecycle.RequestLifeCycleStatus;
 import eu.bbmri.eric.csit.service.negotiator.lifecycle.util.LifeCycleRequestStatusStatus;
@@ -123,7 +124,7 @@ public class ApplicationBean implements Serializable {
 
     public String updateLifecycleStatusProblem_20220124() {
         try (Config config = ConfigFactory.get()) {
-            HashSet<Integer> queryIds = DbUtil.getQueriesWithStatusError_20220124(config);
+            HashSet<Integer> queryIds = DbUtilQuery.getQueriesWithStatusError_20220124(config);
             for(Integer queryId : queryIds) {
                 System.err.println("queryId: " + queryId);
                 try {
@@ -174,7 +175,7 @@ public class ApplicationBean implements Serializable {
                     json = json.replaceAll("https://directory.bbmri-eric.eu/menu/main/app-molgenis-app-biobank-explorer/", "https://directory.bbmri-eric.eu/menu/main/app-molgenis-app-biobank-explorer/#/");
                     json = json.replaceAll("https://directory.bbmri-eric.eu/menu/main/dataexplorer", "https://directory.bbmri-eric.eu/menu/main/app-molgenis-app-biobank-explorer/#/");
                     queryRecord.setJsonText(json);
-                    DbUtil.updateQueryRecord(config, queryRecord);
+                    DbUtilQuery.updateQueryRecord(config, queryRecord);
                 }
             }
             config.commit();
