+++ src/main/java/de/samply/bbmri/negotiator/control/dashboard/DashboardBean.java
@@ -3,6 +3,7 @@ package de.samply.bbmri.negotiator.control.dashboard;
 import de.samply.bbmri.negotiator.control.UserBean;
 import de.samply.bbmri.negotiator.db.util.DbUtil;
 import de.samply.bbmri.negotiator.jooq.tables.records.QueryRecord;
+import eu.bbmri.eric.csit.service.negotiator.database.DbUtilQuery;
 
 import javax.annotation.PostConstruct;
 import javax.faces.bean.ManagedBean;
@@ -32,8 +33,8 @@ public class DashboardBean implements Serializable {
 
     private void collectDataStatistik() {
         queriesInitialized = DbUtil.getNumberOfInitializedQueries();
-        queryRecords = DbUtil.getNumberOfQueries();
-        requestLast7days = DbUtil.getNumberOfQueriesLast7Days();
+        queryRecords = DbUtilQuery.getNumberOfQueries();
+        requestLast7days = DbUtilQuery.getNumberOfQueriesLast7Days();
         requestLineGraph = DbUtil.getDataForDashboardRequestLineGraph();
         humanReadableStatisticsData = DbUtil.getHumanReadableStatistics();
     }
