+++ src/main/java/de/samply/bbmri/negotiator/db/util/DbUtil.java
@@ -102,158 +102,6 @@ public class DbUtil {
         return target;
     }
 
-    /**
-     * Creates a new query from the given arguments.
-     * @param title title of the query
-     * @param text description of the query
-     * @param jsonText the structured data from the directory
-     * @param researcherId the researcher ID that created the query
-     * @return
-     * @throws SQLException
-     */
-    public static QueryRecord saveQuery(Config config, String title,
-                                        String text, String requestDescription, String jsonText, String ethicsVote, int researcherId,
-                                        String nToken,
-                                        Boolean validQuery, String researcher_name, String researcher_email, String researcher_organization,
-                                        Boolean testRequest) throws SQLException, IOException {
-        QueryRecord queryRecord = config.dsl().newRecord(Tables.QUERY);
-
-        if(nToken == null || nToken.isEmpty()) {
-            nToken = UUID.randomUUID().toString().replace("-", "");
-        }
-
-        queryRecord.setJsonText(jsonText);
-        queryRecord.setQueryCreationTime(new Timestamp(new Date().getTime()));
-        queryRecord.setText(text);
-        queryRecord.setRequestDescription(requestDescription);
-        queryRecord.setTitle(title);
-        queryRecord.setEthicsVote(ethicsVote);
-        queryRecord.setResearcherId(researcherId);
-        queryRecord.setNegotiatorToken(nToken);
-        queryRecord.setNumAttachments(0);
-        queryRecord.setValidQuery(validQuery);
-        queryRecord.setResearcherName(researcher_name);
-        queryRecord.setResearcherEmail(researcher_email);
-        queryRecord.setResearcherOrganization(researcher_organization);
-        queryRecord.setTestRequest(testRequest);
-        queryRecord.store();
-
-        /**
-         * Add the relationship between query and collection.
-         */
-        QueryDTO queryDTO = Directory.getQueryDTO(jsonText);
-        for(QuerySearchDTO querySearchDTO : queryDTO.getSearchQueries()) {
-            ListOfDirectories listOfDirectories = DbUtilListOfDirectories.getDirectoryByUrl(config, querySearchDTO.getUrl());
-
-            if (NegotiatorConfig.get().getNegotiator().getDevelopment().isFakeDirectoryCollections()
-                    && NegotiatorConfig.get().getNegotiator().getDevelopment().getCollectionList() != null) {
-                logger.info("Faking collections from the directory.");
-                for (String collection : NegotiatorConfig.get().getNegotiator().getDevelopment().getCollectionList()) {
-                    CollectionRecord dbCollection = DbUtilCollection.getCollection(config, collection, listOfDirectories.getId());
-
-                    if (dbCollection != null) {
-                        addQueryToCollection(config, queryRecord.getId(), dbCollection.getId());
-                    }
-                }
-            } else {
-
-                for (CollectionDTO collection : querySearchDTO.getCollections()) {
-                    CollectionRecord dbCollection = DbUtilCollection.getCollection(config, collection.getCollectionID(), listOfDirectories.getId());
-
-                    if (dbCollection != null) {
-                        addQueryToCollection(config, queryRecord.getId(), dbCollection.getId());
-                    }
-                }
-            }
-        }
-
-        config.commit();
-
-        return queryRecord;
-    }
-
-
-    /**
-     * Adds the given collectionId to the given queryId.
-     * No results from a connector will be expected.
-     *
-     * @param config current config
-     * @param queryId the query id which will be associated with a collection
-     * @param collectionId the collection id which will be associated with the query
-     */
-    public static void addQueryToCollection(Config config, Integer queryId, Integer collectionId) {
-        addQueryToCollection(config, queryId, collectionId, false);
-    }
-
-    /**
-     * Adds the given collectionId to the given queryId.
-     *
-     * @param config current config
-     * @param queryId the query id which will be associated with a collection
-     * @param collectionId the collection id which will be associated with the query
-     * @param expectResults if or not to expect results from a (confidential) connector
-     */
-    private static void addQueryToCollection(Config config, Integer queryId, Integer collectionId, Boolean
-            expectResults) {
-        try {
-            QueryCollectionRecord queryCollectionRecord = config.dsl().newRecord(Tables.QUERY_COLLECTION);
-            queryCollectionRecord.setQueryId(queryId);
-            queryCollectionRecord.setCollectionId(collectionId);
-            queryCollectionRecord.setExpectConnectorResult(expectResults);
-            queryCollectionRecord.store();
-        } catch (DataAccessException e) {
-            // we expect a duplicate key value exception here if the entry already exists
-            if(e.getMessage().contains("duplicate key")) {
-                logger.debug("Duplicate key exception caught.");
-            } else {
-                // TODO: localisation issues? future changes might break this, but then the exception is still caught
-                logger.error("The exception is not matching the phrase 'duplicate key'");
-                e.printStackTrace();
-            }
-        }
-    }
-
-    /**
-     * Flags the given OwnerQuery object with the given flag for the given user.
-     * @param config current database connection
-     * @param queryDto the query object
-     * @param flag the flag that will be set
-     * @param userId the current user ID
-     */
-    public static void flagQuery(Config config, OwnerQueryStatsDTO queryDto, Flag flag, int userId) {
-        /**
-         * Do not hardcode SQL statements. They are hard to maintain.
-         * Since jOOQ does not support the onDuplicateKeyUpdate method yet,
-         * simplify the statements so that:
-         *
-         * 1. If there is no current flag, insert one using the FlaggedQueryRecord class.
-         * 2. If the current flag is the same as the given flag, unflag the query, meaning remove the row from the DB
-         * 3. Update the flag to the given flag.
-         *
-         *
-         * Those are not processing heavy SQL statements, IMHO it's fine.
-         */
-        try {
-            if (queryDto.getFlag() == null || queryDto.getFlag() == Flag.UNFLAGGED) {
-                FlaggedQueryRecord newFlag = config.dsl().newRecord(Tables.FLAGGED_QUERY);
-                newFlag.setFlag(flag);
-                newFlag.setPersonId(userId);
-                newFlag.setQueryId(queryDto.getQuery().getId());
-
-                newFlag.store();
-            } else if (queryDto.getFlag() == flag) {
-                config.dsl().delete(Tables.FLAGGED_QUERY).where(Tables.FLAGGED_QUERY.QUERY_ID.eq(queryDto.getQuery().getId()))
-                        .and(Tables.FLAGGED_QUERY.PERSON_ID.equal(userId)).execute();
-            } else {
-                config.dsl().update(Tables.FLAGGED_QUERY).set(Tables.FLAGGED_QUERY.FLAG, flag)
-                        .where(Tables.FLAGGED_QUERY.PERSON_ID.eq(userId))
-                        .and(Tables.FLAGGED_QUERY.QUERY_ID.eq(queryDto.getQuery().getId()))
-                        .execute();
-            }
-        } catch (Exception e) {
-            System.err.println("ERROR: flagQuery(Config, OwnerQueryStatsDTO, Flag, int)");
-        }
-    }
 
     /**
      * Returns a list of Biobanker's id's who made the sample offers for a given query.
@@ -371,55 +219,6 @@ public class DbUtil {
     }
 
 
-    /**
-     * Gets a list of all the queries from the database
-     * @param config    DB access handle
-     * @return List<QueryRecord> list of query record objects
-     */
-    public static List<QueryRecord> getQueries(Config config, boolean filterTestRequests){
-        Result<Record> result = null;
-        if(filterTestRequests) {
-            result = config.dsl()
-                    .select(getFields(Tables.QUERY))
-                    .from(Tables.QUERY)
-                    .where(Tables.QUERY.TEST_REQUEST.eq(false))
-                    .orderBy(Tables.QUERY.ID.asc()).fetch();
-        } else {
-            result = config.dsl()
-                    .select(getFields(Tables.QUERY))
-                    .from(Tables.QUERY)
-                    .orderBy(Tables.QUERY.ID.asc()).fetch();
-        }
-
-        // TODO: QueryRecord Mapping is not working for requestDescription, ethicsVote and negotiationStartedTime
-        // for this malual Mapping
-        // List<QueryRecord> queries = config.map(result, QueryRecord.class);
-        // Workaround
-        List<QueryRecord> queries = new ArrayList<QueryRecord>();
-        for(Record record : result) {
-            QueryRecord queryRecord = new QueryRecord();
-            queryRecord.setId((Integer)record.getValue("id"));
-            queryRecord.setTitle((String)record.getValue("title"));
-            queryRecord.setText((String)record.getValue("text"));
-            queryRecord.setQueryXml((String)record.getValue("query_xml"));
-            queryRecord.setQueryCreationTime((Timestamp)record.getValue("query_creation_time"));
-            queryRecord.setResearcherId((Integer)record.getValue("researcher_id"));
-            queryRecord.setJsonText((String)record.getValue("json_text"));
-            queryRecord.setNumAttachments((Integer)record.getValue("num_attachments"));
-            queryRecord.setNegotiatorToken((String)record.getValue("negotiator_token"));
-            queryRecord.setValidQuery((Boolean)record.getValue("valid_query"));
-            queryRecord.setRequestDescription((String)record.getValue("request_description"));
-            queryRecord.setEthicsVote((String)record.getValue("ethics_vote"));
-            queryRecord.setNegotiationStartedTime((Timestamp)record.getValue("negotiation_started_time"));
-            queryRecord.setResearcherName((String)record.getValue("researcher_name"));
-            queryRecord.setResearcherEmail((String)record.getValue("researcher_email"));
-            queryRecord.setResearcherOrganization((String)record.getValue("researcher_organization"));
-            queryRecord.setTestRequest((Boolean)record.getValue("test_request"));
-            queries.add(queryRecord);
-        }
-        return queries;
-    }
-
     public static String getFullListForAPI(Config config) {
         ResultQuery<Record> resultQuery = config.dsl().resultQuery("SELECT CAST(array_to_json(array_agg(row_to_json(jsond))) AS varchar) AS directories FROM ( " +
                 "SELECT json_build_object('name', d2.name, 'url', d2.url, 'description', d2.description, 'Biobanks', " +
@@ -443,24 +242,6 @@ public class DbUtil {
     }
 
     /**
-     * Check if the query exists in our system
-     * @param config    DB access handle
-     * @return Query query object
-     */
-    public static de.samply.bbmri.negotiator.jooq.tables.pojos.Query checkIfQueryExists(Config config, int queryId){
-        Result<Record> record = config.dsl()
-                .select(getFields(Tables.QUERY))
-                .from(Tables.QUERY)
-                .where(Tables.QUERY.ID.eq(queryId)).fetch();
-
-        if(record.isEmpty())
-            return null;
-
-        de.samply.bbmri.negotiator.jooq.tables.pojos.Query query = config.map(record.get(0), de.samply.bbmri.negotiator.jooq.tables.pojos.Query.class);
-        return query;
-    }
-
-    /**
      * Check if there are queries expecting results from this collection
      * @param config    DB access handle
      * @param collectionId    unique id of collection
@@ -476,39 +257,6 @@ public class DbUtil {
 
         List<QueryCollection> queryCollectionList = config.map(result, QueryCollection.class);
         return queryCollectionList;
-    }
-
-    /**
-     * A confidential biobanker decides to participate in a query, so we have to add all his collections
-     * to the query_collection table.
-     * If the entry already exists, update the entry to expect results from the connector
-     *
-     * @param config    DB handle
-     * @param queryId   the query ID
-     * @param collections   the collections of the user
-     */
-    public static void participateInQueryAndExpectResults(Config config, int queryId, List<Collection> collections) {
-        if(collections == null || collections.isEmpty())
-            return;
-
-        try {
-            for(Collection collection: collections) {
-                // if already there, update the expect result flag
-                int changedEntry = config.dsl().update(Tables.QUERY_COLLECTION)
-                        .set(Tables.QUERY_COLLECTION.EXPECT_CONNECTOR_RESULT, true)
-                        .where(Tables.QUERY_COLLECTION.QUERY_ID.eq(queryId))
-                        .and(Tables.QUERY_COLLECTION.COLLECTION_ID.eq(collection.getId()))
-                        .execute();
-
-                if(changedEntry == 0) {
-                    addQueryToCollection(config, queryId, collection.getId(), true);
-                }
-            }
-
-            config.commit();
-        } catch(SQLException e) {
-            e.printStackTrace();
-        }
     }
 
     /**
@@ -796,48 +544,6 @@ public class DbUtil {
         return 0;
     }
 
-    public static String getNumberOfQueriesLast7Days() {
-        try (Config config = ConfigFactory.get()) {
-            Result<Record> fetch = config.dsl().resultQuery("SELECT COUNT(*) FROM public.query WHERE query_creation_time > current_date - interval '7 days';").fetch();
-            int querys_created = 0;
-            for(Record record : fetch) {
-                querys_created = Integer.parseInt(record.getValue(0).toString());
-            }
-            Result<Record> fetch_json_query = config.dsl().resultQuery("SELECT COUNT(*) FROM public.json_query WHERE query_create_time > current_date - interval '7 days';").fetch();
-            int querys_initialized = 0;
-            for(Record record : fetch_json_query) {
-                querys_initialized = Integer.parseInt(record.getValue(0).toString());
-            }
-            return querys_created + "/" + querys_initialized;
-        } catch (SQLException e) {
-            System.err.println("ERROR getting open Request Status.");
-            e.printStackTrace();
-        }
-        return 0 + "/" + 0;
-    }
-
-    public static List<QueryRecord> getNumberOfQueries() {
-        List<QueryRecord> returnList = new ArrayList();
-        try (Config config = ConfigFactory.get()) {
-            return config.dsl().selectFrom(Tables.QUERY).orderBy(Tables.QUERY.QUERY_CREATION_TIME).fetch();
-        } catch (SQLException e) {
-            System.err.println("ERROR getting open Request Status.");
-            e.printStackTrace();
-        }
-        return returnList;
-    }
-
-    public static List<QueryRecord> getNumberOfQueriesAssociatedWithNetwork(Config config, Integer networkId) {
-        Result<Record> record = config.dsl().select(getFields(Tables.QUERY))
-                .from(Tables.QUERY)
-                .join(Tables.QUERY_COLLECTION).on(Tables.QUERY_COLLECTION.QUERY_ID.eq(Tables.QUERY.ID))
-                .join(Tables.NETWORK_COLLECTION_LINK).on(Tables.NETWORK_COLLECTION_LINK.COLLECTION_ID.eq(Tables.QUERY_COLLECTION.COLLECTION_ID))
-                .where(Tables.NETWORK_COLLECTION_LINK.NETWORK_ID.eq(networkId))
-                .orderBy(Tables.QUERY.QUERY_CREATION_TIME)
-                .fetch();
-        return config.map(record, QueryRecord.class);
-    }
-
     public static String getRequestsForNetworkAsJson(Config config, Integer networkId) {
         ResultQuery<Record> resultQuery = config.dsl().resultQuery("SELECT CAST(array_to_json(array_agg(row_to_json(jsonc))) AS varchar) FROM ( " +
                 "SELECT q.id query_id, q.title query_title, b.id biobank_id, b.name biobank_name, b.directory_id biobank_directory_id, " +
@@ -1004,25 +710,6 @@ public class DbUtil {
                 .fetch();
 
         return config.map(result, QueryRecord.class);
-    }
-
-    public static void updateQueryRecord(Config config, QueryRecord queryRecord) {
-        queryRecord.update();
-    }
-
-    public static HashSet<Integer> getQueriesWithStatusError_20220124(Config config) {
-        HashSet<Integer> queryIds = new HashSet<>();
-        ResultQuery<Record> resultQuery = config.dsl().resultQuery("SELECT query_id\n" +
-                "FROM (SELECT query_id, STRING_AGG(status, ',') status_string FROM request_status GROUP BY query_id) AS sub\n" +
-                " WHERE (status_string ILIKE '%created%' AND status_string NOT ILIKE '%under_review%') AND \n" +
-                " (status_string ILIKE '%created%' AND status_string NOT ILIKE '%abandoned%');");
-        Result<Record> result = resultQuery.fetch();
-        if(!result.isEmpty()) {
-            for (Record record : result) {
-                queryIds.add(record.getValue(0, Integer.class));
-            }
-        }
-        return queryIds;
     }
 
     public static void updateCollectionNetworkLinks(Config config, DirectoryCollection directoryCollection, int listOfDirectoryId, int collectionId) {
